requirements.txt
python-whois==0.9.5